Add files via upload
Over the years, the company has
collected details and gathered a lot of
information about individuals. The
management wants to build an intelligent
system from the data to determine the
presence or absence of smoking in a person
through bio-signals. Given a person’s
information, build a machine learning model
that can classify the presence or absence of
smoking.